github: rework Docker build, and publish new images
This reworks the `Dockerfile` to now build multiple images, including
one for end users to run their own copy of Glean, and an updated version
of the React demo database. It includes x86_64 and aarch64 Linux images
using public GHA runners, which probably covers 98% of all realistic
uses. (Using GHA runners means this workflow is trivially usable for
public forks, too.)

There is a new `bindist` target used by the Dockerfile that creates a
directory of needed binaries and libraries, and properly packages them
for usage in the container.

There are now three containers published from the `Dockerfile`, all
based on Ubuntu 24.04:

- `glean/client`, which only contains the `glean` command, then
- `glean/server`, including `glean-server`, then
- `glean/demo`, including a precanned `glean-hyperlink` + react DB

These containers are all size-optimized; the final images have no
development tools installed and the absolute minimum necessary installed
packages. For example, the new x86_64 `glean/demo` is only 486MB
uncompressed, compared to the 2.4GB of the previous demo version. The
`client` and `server` images are 251MB and 300MB, respectively.

The `client` image should probably include `gen-schema` so that end
users can generate code from their own non-upstream `.angle` files, but
that can be added in a future patch.

In theory, this new container could also be used to replace most of the
junk in `ci.yml` in order to repeat ourselves less, but it would need
another image in order to add more things so that the test suite can be
run. That can happen in a future patch.

Signed-off-by: Austin Seipp <aseipp@pobox.com>