Fix element generators of bytes
Summary:
In a D68322262 I fixed how we write a byte to memory. But there was still a problem remaining for how we read a byte when we use the elements construct. We were still reading using the packed representation. This diff fixes that by reading using the fixed representation instead.
I've had to add a new bytecode instruction to fix it.

There's also a fix for the `all` construct when generating bytes.

Reviewed By: simonmar

Differential Revision: D68636014

fbshipit-source-id: 208ad0c697f9b5d15cb6e2474803f91c8c18bb78