Overload the elements syntax to work for both arrays and sets (take 2)
Summary:
The biggest lesson from with migrating arrays to sets is having to change all uses of the elements syntax from X[..] to elements X. This also caused a lot of breakage for a variety of queries in production. The right way to fix this is to just overload the syntax for arrays to also work for sets. That's what this diff does.

I had a previous crack at this in D68559811 using a form of constraint machinery. This diff uses a more straightforward unification mechanism and is simpler.

Reviewed By: simonmar

Differential Revision: D68624547

fbshipit-source-id: aa42f879e54f94c434b7c07433c1be27fc6000d9