Move commitOwnership under the write lock
Summary:
In D52733037 we moved commit outside the write lock, but this also
moved `addDefineOwnership` outside the write lock, which isn't safe
because it uses `db->usets_` which needs to be protected by the write
lock.

This is very probably the cause of various weird ownership-related
crashes we've seen.

I didn't want to completely revert D52733037 because that's a pretty
big perf win for writing. I verified that it's still a win by
repeating the test plan from D52733037.

Reviewed By: josefs, iamirzhan

Differential Revision: D69048830

fbshipit-source-id: ba4ad242f1ebcdec86fdb6e36d6c88ef17eadcdc