add called function in buck.TargetCall predicate
Summary:
`buck.TargetCall` contains the arguments and span of a Buck build rule. It is used currently in `fbsource.bzl` to represent build rules.

This diff adds to this predicate the call expression used by the rule, as well as a derived predicate so we can find efficiently all the call sites for a given call expression.

Users want to get all the target calls for a given macro.`buck.DestinationUses` provides similar information, but only allows to find the spans of references to the macro definition rather than the target call.

Note: fbs-validate-www-changes  is expected to fail.  glean/schema/sync script now generates www code directly. This shouldn't be an issue as we expect sync-cf workflow to deploy the new schema to configerator before www is deployed.

Reviewed By: malanka

Differential Revision: D69047083

fbshipit-source-id: 698a4be3a6f59499188fe24ba92d490d53d2c445