Add batch descriptor to server write queue
Summary:
This is the first part of `enqueueBatchDescriptor` implementation. The batch descriptor should be added to the same writeQueue with other batches.

The only difference is on this method the server will not check the queue size and not throw a retry. We assume the batchDescriptor size is 0 (as no real batches are added to the queue). Later the batch will be downloaded (we might need to control the total size of downloaded but not written batches, but it's anyway bounded by threads count, so should be ok for now)

When batch descriptor is dequeued, the writing thread first will download the batch and then write it. The downloading part is not implemented yet.

Reviewed By: pepeiborra

Differential Revision: D69668661

fbshipit-source-id: 80fa32913af522f3576ee8310b75c6b17724c5a7