Make scope.thrift annotations "magic"
Summary: We can't use structured annotations in this file because of circular dependencies, and we don't want to keep the unstructured annotations around. Inject the annotations in the mutator where they will not be picked up by the codemod and we can eventually inject them as structured once the bundling work is complete.

Reviewed By: yoney

Differential Revision: D69817266

fbshipit-source-id: b39c204b2633ee4db9971f490b3b24ad342fe967