Make --replace-all work for OSS
Updating test output is a bit of a pain with Cabal,  because you have to
look up the test root (which is already set by the test) and pass it as
an argument to `--replace`. With Buck this is all handled by `-c
glean.test=replace` but we don't have that for Cabal.

This makes `--replace-all` just use the root provided by the test in the
OSS build.

To update test output, e.g.

```
cabal run glean-snapshot-typescript -- --replace-all
```