New Haskell Indexer
Redesigned the schema and rewrote the indexer. Compared with the
previous indexer:

* this captures a lot more xrefs (e.g. local variables)
* it has more information (distinguishes functions/classes/constructors etc.)
* it is much simpler and probably faster, because it doesn't go via hiedb, it reads .hie files directly.
* it is probably more correct, I fixed a lot of things.

The schema is carefully designed so that a Name uniquely identifies an
entity and corresponds fairly closely to GHC's Name, including
OccName. The main difference is we don't store Uniques, instead we
distinguish local Names by including their ByteSpan.

There are a couple of snapshot tests, one for the plain indexer and one
for the codemarkup layer.

Not done yet:

* Removing all the old stuff

* more Glass integration, e.g. Symbol ID

* we can extract types from the .hie file too, and provide type hovers
  in Glass. That wouldn't be too hard.

* extracting more structure so that we can reconstruct data/class decls
  should be possible but it's not straightforward using .hie. I'm still
  thinking about how best to do that.