Regenerate the scip index for the scip/rust tests
Summary:
Regenerating the index.scip index for rust tests with a newer version of rust-analyzer to have the latest SCIP fields, which I plan to use in a future diff.

The regeneration method is also added in index.README.md

This changes a lot of test file outputs. Most of these query all the facts of a certain predicate. I'm wondering if querying only a few facts would make the tests more relevant and less brittle. Otherwise not sure what the purpose of these tests is.

Reviewed By: donsbot

Differential Revision: D73101667

fbshipit-source-id: 291ff644b28c0283be5d7d4d62b09340ab7ebbf2