add a new predicate for the mangled name hash
Summary:
I've been trying to ingest the strobelight qualified names into glean (which are rather huge) for frame matching.

dmpolukhin suggested an alternative approach to try and get hashes of mangled names into the C++ DB (which should not be a lot of storage)  and then we can use these for a 1:1 match on the strobelight data (with no need for more complex imprecise matching)

I'm adding the predicate to the DB so I can try it out and measure the impact on the size of the DB. If the DB is too large etc, ill just clean it up.

Reviewed By: iamirzhan

Differential Revision: D73593682

fbshipit-source-id: 283f1bd8e090428a8cd2d46b47da7dd0354d9dc6