Add source span to all type references in angle AST
Summary:
Adding srcSpan to the Angle's Type def

**before**
pref and tref for Predicate and Type references
```Type = Type_ pref tref```
**after**
```Type = Type_ st pref tref```

Thus, any identifier of a type/predicate will have a span (st), so we can get the XRefs with spans.

**Tradeoff:**
* Introduced another type variable st, such that some types will have *two* type variables for spans (s and st). This is not optimal imo, but it makes the size of this diff smaller.
  * This allows RTS Type to have ```st = ()``` whilst ```s =SrcSpan```. If we didn't do this we'd need to refactor all RTS types too (follow the ToRtsType, TypecheckState).
I'd suggest to unify this in a follow up diff (if we want to)

https://www.internalfb.com/code/fbsource/[ba4e651685d3066b481b1241790a3449b8d37993]/fbcode/glean/hs/Glean/RTS/Types.hs?lines=111
https://www.internalfb.com/code/fbsource/[9de842f2288336dd3a0c1b1c9b679e2b70b65504]/fbcode/glean/db/Glean/Query/Typecheck/Monad.hs?lines=56

**Affects codegen**
Note that diff affects the schema code gen for Angle->Haskell.

Reviewed By: pepeiborra

Differential Revision: D73190439

fbshipit-source-id: 0887c0d9bb232c7e341b6d928cd6dcb8dbbae4e9