Fix bug when outputting JSON
Summary:
The JSON output by the scip indexer was subtly bogus. There would be weird occurrences of unparsable pieces of JSON, and a recurring issue was that the terminal `]` would be issued in the middle of the file instead of at the end.

I rewrote the logic writing a list of JSON to file to use `Data.ByteString.Lazy.hPut fileHandle` instead of `Data.ByteString.Lazy.appendFile filePath`, and now it works. The documentation of the latter is very succinct, and I suspect it's non-blocking and interspersing things in the file.

Reviewed By: donsbot

Differential Revision: D73858723

fbshipit-source-id: 240c1a52e5970fbe62ae8415252b1436ec74f6c5