[new-hs-indexer] tag refs with import/export/code
It's useful to be able to know whether a ref is in an import or export,
e.g. for dead code detection, so let's tag refs with an enum to say what
kind of ref it is.

I changed field names so that the schema change shouldn't trigger a
validation failure (I hope).