add method USRToDefiniton
Summary:
High-level goal: use glass as a remote index store for sourcekit LSP. To that end, this diff adds thrift method usrToDefinition which takes a sourcekit USR (unique symbol identifier in clang/swift) and return the symbol definition.

This method is based on a glean query similar to the one in glean/lang/swift/tests/regression/usrToDef.query, i.e.
```
    SE = { swift = { defn = { symbol = "myUSR" } } };
    codemarkup.scip.ScipEntityLocation { SE, L };
    {{ scip = SE }:code.Entity, L}
```

Reviewed By: donsbot

Differential Revision: D74321433

fbshipit-source-id: 4d0e6e5d40b54fd7df14aece2d25d086400e1e74