add schema for stack parents/children
Summary:
we're going to merge the performance tab with the current strobelight tab in Codehub. the main missing feature is this hovercard.

adding the schema for it. the data seems reasonable in size as we only cover the top few parents/children (current dataset strobelight tab uses is  ~730 MB which covers ~33 services)

 {F1978034775}

Reviewed By: malanka

Differential Revision: D74883752

fbshipit-source-id: bce6524cc6912d66e58e66ff39d44baadf960ec7