use sharding parameter to shard output
Summary:
This diff adds a sharding algorithm for the SCIP JSON schema.

It's motivated by the fact that we can't write crazy huge files into Glean, it invariably times out. But we also can't write a single section of a JSON file into Glean, because the foreign keys used to create the graph structure are ephemeral, and Glean rewrites the IDs into its own ID form.

This means we have to shard complete subgraphs. This means nodes may be written multiple times, but if Glean times out on writes, we can just retry. This also allows us to throttle and back off the amount of data we write.

To shard complete subgraphs, I categorized node types from the SCIP schema into source nodes, sink nodes, and all other nodes. The entry point into the graph is source nodes (nodes with only edges going out). We iterate over each source node and do a full traversal of the graph that is reachable from that source node. We continue iterating until we have a subgraph of a certain size. Then we reset the list of visited nodes (since each shard must contain a complete subgraph) and create a new subgraph.

Sink nodes are the leaves, so we stop the traversal when we visit those nodes.

It's annoying to write this traversal in one single function, but the alternatives were even more annoying (structs or other functions with lifetime specifiers everywhere), though I'm open to other approaches to writing this so it's not a giant function. I also didn't want to copy multiple gigabytes of data.

Another possible approach is to iterate over graph edges, though you would still have to traverse a complete "chain" of nodes

**Tests are coming, I just wanted to get this out there to unblock Java indexing**

Differential Revision: D72428666

fbshipit-source-id: de7624dcde87a880174f003e5f748adc6eb397fa