also index USR hashes of objc things
Summary:
The cxx indexer uses a predicate USRToDefinition. This is a mapping from USR hashes to definition. USR are unique symbol indentifiers used in cxx toolchain internals. The USR hashes are computed by the indexer itself. However, so far it didn't bother with USR hashes of Objc symbols. This diff changes that.

This will by use to implement swift->objc cross language navigation in the next diff.

Reviewed By: iamirzhan, nhawkes

Differential Revision: D76419553

fbshipit-source-id: 0860346025fe0f04db20d115024c871a061b6c22