Install the schema and don't require a `--schema` flag (#545)
Summary:
I wanted to make the OSS build install the schema under `$datadir` and to look for it there by default. This meant that the default schema location is an `IO` operation (because `getDataDir` is in `IO`), which needed some refactoring. While I was there I noticed that the way schema flags work was a bit of a mess, and that we had no way to specify the schema location in the server config which is weird when everything else is configurable that way.

So the main changes here are:
* OSS builds install the schema in `$datadir`, and binaries no longer require a `--schema` flag if you just want the default schema.
* New type `SchemaLocation` in `server_config.thrift`
* New field `optional SchemaLocation schema_location` in `ServerConfig`
* Change from `cfgSchemaSource :: ThriftSource SchemaIndex` to `cfgSchemaLocation :: SchemaLocation` in `Config`.

and various consequential changes as a result.

NOTE: when importing this change you will need to update the internal `Glean.DefaultConfigs` to add `defaultSchemalocation`, should be something like

```
defaultSchemaLocation :: SchemaLocation
defaultSchemaLocation = SchemaLocation_indexconfig schemaConfigPath
```

Pull Request resolved: https://github.com/facebookincubator/Glean/pull/545

Test Plan:
Imported from GitHub, without a `Test Plan:` line.

Rollback Plan:

Reviewed By: simonhollis

Differential Revision: D76969691

Pulled By: jjuliamolin

fbshipit-source-id: 937201f49ae713c879ab8d4db6ea0f86dd313279