Fix type checking of `[..]` and `elements`
Bug #1: `X[..]` behaved differently depending on whether we were inferring or
checking types. When checking types we forced `X` to have an array type,
whereas when inferring types we allowed `X` to be either an array or a
set. This led to some very confusing behaviour.

Bug #2: the type checking of `elements X` when expecting a predicate
type was wrong, due to the ordering of pattern matches in
`typecheckPattern`. This also led to some very confusing behavoiur.