Update Rust codegen script to account for bugs that we might see
Summary:
Currently the codegen works perfectly fine for the files we need, but if we run codegen on additional files we notice there are bugs and breakages. We should really fix this in case the schema files we do need change in such a way that breaks the codegen.

Here are some fixes:
- Add `safe` function which fixes the name of a name in case it matches some reserved name.
- Fix the `builtin.Unit` to be `builtin::Unit`
- Because of the constructor code we added below this diff, it is possible we generate duplicate definitions for types that are used in predicates. To fix this, when we generate the constructor I now ensure we don't `push` the new type to the list of types of construct. In all other cases, we do `push` the type. This ensures it is only generated once.

Reviewed By: nhawkes

Differential Revision: D77307704

fbshipit-source-id: 3672fd50a04967764bbd36c92dbeae2fa5b4ed1f