Merge improvements 2/2: Adding read of schema from JSON
Summary:
`glean merge` officially supports supplying both databases to be merged and JSON files + and inventory file to be merged.

However, if supplying raw JSONs, there's no `dbSchema` available, and the merge fails. This diff handles the raw JSON case by going off to configerator and reading the schema definitions there.

It relies on a `schema_id` field being present in the JSONs to lookup the correct schema.

Reviewed By: iamirzhan

Differential Revision: D73030289

fbshipit-source-id: 9c6ea68acfaf383d1b7f865342f2cc6f1aff5556