Introduce test for JsonServer
Summary:
This diff introduces a test for the `JsonServer` class in the `glean/client/swift` directory. The test, named `JsonServerTest`, is a C++ unit test that verifies the functionality of the `JsonServer` class.

Note: there is only one test added and that can be considered as a test stub because it tests the echo functionality that will be replace in D78900349

Reviewed By: dmpolukhin

Differential Revision: D78892919

fbshipit-source-id: 7dce3971af87f2a0fd5c91951df339ed400aff14