Communication with Glass via ServiceRouter
Summary:
This diff updates the BUCK file for the Glean Glass client in Swift, adding dependencies for ServiceRouter and folly/coro. It also modifies the GlassAccess class to use ServiceRouter for communication with the Glass service, and implements the usrToDefinition method.

With the diff we get a functional Swift Glass Client application that can be run on DevServers and OnDemand servers. It's important to note that it cannot be built on local (Darwin) and the required changes will be implemented later in subsequent diffs

Reviewed By: dmpolukhin

Differential Revision: D78901802

fbshipit-source-id: b80b8c67dd1e09c3e26bbcf56f6ee92ffa71e0cd