IGlassAccess introduced
Summary:
This diff introduces the `IGlassAccess` interface and its implementation, `GlassAccess`, in the `glean/client/swift` directory. It also updates the `JsonServer` class to use the new `IGlassAccess` interface.

The `IGlassAccess` interface defines a single method, `usrToDefinition`, which takes a `std::string` parameter and returns an `std::optional<protocol::LocationList>`. This method will be used to access Glass data.

The `GlassAccess` class implements the `IGlassAccess` interface and provides a concrete implementation for the `usrToDefinition` method.

The `JsonServer` class has been updated to use the new `IGlassAccess` interface. It now has a `setGlassAccess` method to set a `std::unique_ptr<IGlassAccess>` instance. The `start` and `stop` methods have also been updated to use the `IGlassAccess` interface.

Additionally, a mock implementation of `IGlassAccess`, `GlassAccessMock`, has been added to the `test` directory. This mock implementation is used in the updated `JsonServerTest` class to test the `JsonServer` class with a mock `IGlassAccess` instance.

The `BUCK` file has been updated to include the new `glass_access` library, which includes the `GlassAccess` and `IGlassAccess` headers. The `json_server` library now depends on the `glass_access` library. The `GlassSwiftClient` binary has also been updated to depend on the `glass_access` library.

Reviewed By: dmpolukhin

Differential Revision: D78892920

fbshipit-source-id: 63d87bcda06ec8a4349e4bc7ed64e3cafa152a5c