Glass connection warm-up
Summary:
The diff implement a connection warm-up mechanism in the GlassAccess constructor that sends the `getStatus` request to establish the connection. The diff also logs the timing.

Note: based on the tests with binary built with mode/opt flag file - the initialisation takes ~1.5sec. With mode/dev it will take up to 30sec

Reviewed By: Wilfred

Differential Revision: D78978796

fbshipit-source-id: 7f67feb963930a181e0cfaebb372d32364428692