add a custom Swift schema extending the SCIP schema
Summary:
This diff adds a custom Swift schema swift.angle which imports the scip schema. This should allow more flexibility in how we index swift in the future.

The schema contains one predicate for now, SymbolLanguage, matching symbols to languages. This will allow to fix cross language references, which is currently broken. Currently, it relies on symbols starting with `c:objc`, but that's unfortunately not always correct. Some ObjC symbols are of the form `c:@...@objc`. In a future diff, we'll modify the indexer to demangle symbols and detect that pattern to add facts in this new predicate (for now, a vague proof of concept is in D79262434).

Reviewed By: aahanaggarwal

Differential Revision: D79261787

fbshipit-source-id: fbfaaa7e5fce433eee3fed0536baec48f0f19571