add predicate to obtain symbol ID of a member
Summary: Right now in the swift/scip schema, most predicates require a scip.Symbol (i.e. a symbol ID) to be queried. Symbols IDs for classes can be retrieved with scip.DisplayNameSymbol, but it's much harder for members. In addition, it's hard to connect containers and their members. This diff fixes that.

Reviewed By: phlalx

Differential Revision: D79338724

fbshipit-source-id: 37118a3ae7bef7d658d37a012b0595fc73340fda