Amend reference line numbers when revisions do not match
Summary:
# What this diff does

When we request a file in a revision A that has not been indexed, Glean provides the best approximation, revision B.

If the file changed between revisions A and B, the symbol index will contain erronous line locations, which will make CodeHub render the wrong symbol information when hovering.

This diff tries to mitigate this problem by calculating the correct line number for the parts of the file that remained unchanged between both revisions.

# Caveats

- This is a heuristic, if lines were deleted between both revisions, that information is simply not available.

- Symbol target locations were not resolved, they could be pointing to other files with similar problems or refer to lines that don't exist in revision A. I do not think this is worth improving for now.

# Questions

- The motivation is CodeHub, I want to validate this change won't interfere with other services, how can I do so?
- I implemented this change in the location I considered more appropriate (`fetchSymbolsAndAttributesGlean`, before adding dynamic attributes), I'd appreciate feedback on this

 ---
- before:
*Notice: references missing, the ones that show up are wrong*
 {F1980381982}

- after:
*Notice: references and definitions in place*
 {F1980381962}

Reviewed By: kbojarczuk, phlalx, jjuliamolin

Differential Revision: D78435339

fbshipit-source-id: 5778a01e8b536eb91ab86b8358d6716806aa0c3e