add an option to override the binary in default view
Summary:
I'm working on an option to change the binary directly in default view (from the hover card): D79704296

Realizing that this means that we also need to change the default view query in the API to allow overriding the binary selected in case the user wants to change it.

The reason we need to use the default view query and not the "regular" query, is because it has line mapping which is essential to see the data in default view.

Differential Revision: D79704900

fbshipit-source-id: 4d214202c870bc1a41bbfa118700d4272ba80c72