Swallow error when calling pyrefly indexer and remove comments from test python code
Summary:
pyrefly will exit with code 1 if there are any type errors in the code. this fails the indexing run (even though the correct JSON's are produced).

I could add a flag on pyrefly that will exit with code 0 but the pyrefly team didn't seem to be a big fan there.

We want to ignore this failure and just use the JSON. Pyrefly could fail with exit code 2 if there's an infra failure, but in that case the test will still fail if the JSON's it produces are wrong.

Reviewed By: CatherineGasnier

Differential Revision: D80010202

fbshipit-source-id: 69ec2a7a6949ff6a00744472a508397f5b423eb0