add an option in Glass to override the service in default view
Summary:
## Original
Walking along the lines set by D79704900, we would like to add an option to explicilty set which service is being used in default view.

The use case here is actually not for default view directly, but for the data presented in vscode: in VSCode we can't force the user to be based on a certain commit, so we want to allow using the line matching algorithm (which takes place if we are passing the default view flag).

Since we would also like to allow setting the binary and service being used, it makes sense to add options to override these as well.

## Update
Instead of going parameter by parameter, adding a new flag to respect the parameters passed also in the private case of default view.
We will utilize this functionality to respect the selections of the user even if you are not on the same revision as the data. The key is the line matching algorithm.

This will be used for both VSCode and Code Hub, across all parameters (including assembly).

For now, only extending it to service+binary in cost as it's the most straightforward.
Since the binary override is already in use, let's keep compatibility with that functionality in Glass as well.
After this diff is pushed and we update all usages of the new flag, will cleanup the usage of the old flag.

Reviewed By: kbojarczuk

Differential Revision: D80127945

fbshipit-source-id: a19adb90aff9efb2ae6712833f6120b42c400a7a