Async test for Glass Client
Summary:
This diff introduces an asynchronous test for the Glass Client in Swift, verifying its functionality in async mode. The test, `test_usr_to_definition_swift_class_with_delay_async`, sends three requests with different delays to the Glass Client and checks their responses.

To support this new test, the diff also adds the necessary dependencies to the `BUCK` file and includes the required headers in `JsonServer.cpp` and `JsonServer.h`. Additionally, it creates a `CancellableAsyncScope` object in `JsonServer` to manage the asynchronous tasks.

Reviewed By: dmpolukhin

Differential Revision: D80700092

fbshipit-source-id: 130573099fbbdb1f85469beb96f968d2f44dec34