Handle bad working directory gracefully
Summary:
The glass client was throwing an exception inside a constructor if `hg root` didn't succeed, leading to a somewhat unhelpful stacktrace if we use the glass client in the wrong directory.

Instead, get the root directory during startup, and terminate with a clear message if the directory is incorrect.

Reviewed By: CatherineGasnier

Differential Revision: D81327031

fbshipit-source-id: e4c2bcd6572330a67131cbf1d640f8415406f6f1