Add request option that puts pyrefly repo on top of repo mapping if flag is true
Summary:
I want to be able to test the pyrefly repo, and check the navigation performance for instagram against the current python indexer.

The only differences between the 2 repos based on regression tests now is variable assignments, where the references will not be found for both sides of the operator - this should not be very consequential.

Using this gk, we can have metrics to see UPNU changes or changes between groups.

Reviewed By: phlalx

Differential Revision: D80803054

fbshipit-source-id: 04b0e4a9b64fd8f0d7d41742fc5637cc32399ea4