Add regression tests for declaration locations
Summary:
Ad title.
To validate results, I created the notebook N8232705 to analyse the differences between the declaration locations using the output produced in this diff, and the output produced by the original indexer on the pyrefly file (see D83661000)

Differences are due to pyrefly adding information that is missing in the python indexer, or having slightly more accurate location for definitions.

Reviewed By: kbojarczuk

Differential Revision: D80257537

fbshipit-source-id: 27ea0ff796a3f0dcd0a9c1db8af0cc0b9819cec3