Add CommonDependency predicate
Summary:
We currently limit go-to-def within the same "root" to avoid returning definitions outside the branches. However, there are important dependencies that are outside of the directory branches. We plan to fix leveraging pyrefly, but we were asked for quick mitigation for specific directories: `genai/msl/rl`, `genai/llama4x` and `genai/msl/sixlib` which are the ones they care the most.

This predicate will allow us to define common dependencies quickly to mitigate the issue while we work on the long term solution that involves pyrefly.

Note: there was a complaint from a VP in MSL, so it seems to be a high priority bug

Reviewed By: CatherineGasnier

Differential Revision: D84061541

fbshipit-source-id: bb30d0eee9021d7fa9704bf1a4e397a7e3376494