Take isolation directory as an argument
Summary:
In D74321433 we changed the isolation directory used in all buck invocations by the indexer. Buck caches use absolute paths, and buck output directories depend on the isolation directory. For example, with `--isolation-directory foo` the output is written to `buck-out/foo/`.

As a result, buck invocations by the indexer got no benefit from existing buck caches from e.g. sandcastle builds. The intention in D74321433 was to allow the Glean tests to invoke buck, because buck does not allow recursively calling itself unless the isolation directories are different.

Set the isolation directory when the Glean indexer tests run, but not otherwise.

Reviewed By: CatherineGasnier

Differential Revision: D84149170

fbshipit-source-id: 04369b6ba594d9ba3ae9f92d06291292c6baef4d