Hacky\Normalize TS filepaths in scrip/angle converter
Summary:
TypeScript type service provides us with relative filepaths (bug that they're not going to fix quickly).

It causes issues with gotodef in codehub browse because browser prevents us from navigating to URLs with relative paths for security reasons.

We've limited to TS files only since we don't want to apply this normalization for backends that properly return paths without ../.

Reviewed By: donsbot

Differential Revision: D84070176

fbshipit-source-id: 55adc2d832f6b0bffdd68de8c624d2df27404f98