test wrong entity
Summary:
The thrift_library in the following code is indexed a buck build rule.

```
def bound_variable(MY_CONST):
    if has_thrift_model:
        thrift_library(
            name = "model",
            languages = ["py-deprecated"],
            thrift_py_options = "json",
            thrift_srcs = {
                "model.thrift": [],
            },
        )
    return MY_CONST
```

The indexer is over-eager. Adding the test case first to exhibit the behavior, fixing in next diff.

This isn't correct as this isn't a build_rule, there can't be  valid xrefs to it.

This was reported in this [comment](https://fb.workplace.com/groups/factsaboutcode/permalink/2674349616247603/).

Reviewed By: malanka

Differential Revision: D85055288

fbshipit-source-id: 13552da41b95e6ad517405f61617ae387f897779