Index new predicates USRSymbol and SymbolUSR
Summary:
The previous diff adds predicate to match scip.Symbol to a USR (and vice versa).
Currently, for swift, the scip.Symbol is actually the USR instead of following the format required for SCIP symbols (see third-party/scip/scip.proto).
This gives rise to USR-like things in CodeHub (e.g. in symbol search), because our SCIP indexer actually demangles SCIP symbols. This stack of diffs aims at fixing that.
This diff will populate the predicates added in the previous diff. For now, since Symbols are actually USRs, it matches USRs to USRs, but we'll later change symbols to be proper SCIP symbols instead.

Reviewed By: Wilfred

Differential Revision: D85145184

fbshipit-source-id: 2ad8970012dd9232f5b29175b9ffc047e8ac6473