extend argument values
Summary:
extending the fbthrift.angle schema to so we don't miss any ArgumentValues when indexing calls. Similar to what is done in Hack. https://www.internalfb.com/code/fbsource/[3a9765d46d55]/fbcode/glean/schema/source/hack.angle?lines=525

Whe value are name, we can resolve the name to a definition. When arguments can't be directly represented in Angle, we use the other case and use a span so that client can inspect manually the argument if needed.

Reviewed By: jjuliamolin

Differential Revision: D85146402

fbshipit-source-id: f38ddc2dd1cb002b279fb948ab902b822f6b0b7e