remove the scip-to-glean haskell indexer
Summary:
To convert scip indexes to glean, we have 2 indexers, one written in Rust, the other written in Haskell. The haskell one isn't used anywhere, and is less performant. Maintaining two indexers is pointless. Therefore I'm removing the Haskell indexer

Most of the changes in this diff are in Haskell wrappers used for tests.

Reviewed By: kbojarczuk

Differential Revision: D85246227

fbshipit-source-id: 784514147a1d5d92f390f2c76886e5bda3989fb3