Regression tests for Glean indexer
Summary:
# Context

Indexing of Erlang files to produce Glean facts is supported by the [ELP project](https://github.com/whatsapp/erlang-language-platform) via the `elp glean` command.

# Problem

We plan to perform several changes to the Erlang indexer. For example, symbols such as variables are incorrectly classified as functions. Indexing also supports two versions (`v2` vs `v1`), where `v1` is not used any longer.

To be able to comfortably update the `glean` component in ELP, we want to introduce snapshot tests, that will increase our confidence when updating the indexer.

# This diff

Here we add snapshot tests for the Erlang indexer, using other languages and the official documentation as reference. Since I'm not familiar with the `glean` codebase, please review this carefully and advice accordingly.

Tests are added both for Erlang and Codemarkup (used for CodeHub).

We also add the Erlang indexer to the list of Glean indexers, so it's possible to use it via:

```
glean index erlang ...
```

Reviewed By: phlalx

Differential Revision: D85324457

fbshipit-source-id: 1ed92ea93b139b90da210750d395d9167d0cead5