Add regression tests for xrefs
Summary:
As title. I analysed the differences running the tests with both indexer on the same files which here N8390972.

The symbols that are missing are the symbols imported with the exported name, but since we always jump now to the original definition we don't need them  (I don't think we need them in the current indexer either).

The differences in location for some of the symbols are improvement instead of regressions. For example, we now report original import instead of intermediate import, and we are more accurate when we report location of types annotations.

Reviewed By: malanka

Differential Revision: D85664629

fbshipit-source-id: 26b453417c5d038bc541a54f833375efc6c698e1