Update codemarkup schema to use SCIP EnclosingRange whenever available as location
Summary:
D84823379 introduced EnclosingRanges for SCIP and populated it for Golang.

This means that for a variety of symbol kinds, like funtion, while the symbol range remains untouched, you can find the enclosing range (lexical scope) by querying this predicate (if it exists).

This diff edits codemarkup, which is the language-agnostic layer that Glass queries.
I've edited `ResolveLocation` and `EntityLocation` which are used to translate between language-specific world (entities) to language agnostic world (location) to try and use the enclosing range if available as the location of the entity.

Reviewed By: kbojarczuk

Differential Revision: D86365726

fbshipit-source-id: 063cce33f8ead90e31282e15eda11ebcc4cd581e