use file mapping when flag is on
Summary:
The logic here is the final step in the chain of our effort to resolve strobelight file paths (which are partial):
1. Resolve the file paths (done)
2. Ingest the full paths (done)
3. Map the resolved paths which are not selected to the key resolved path (this diff)

Basically we select a single resolved path as the "key" path for the data and ingest that. But there could still be other resolved paths that we want to show the data as well.

For example:
- Strobelight file path is: `quic/common/ContiguousCursor.h`
- Resolved to two paths:
```
[
  "fbsource/fbcode/quic/common/ContiguousCursor.h",
  "fbsource/xplat/quic/common/ContiguousCursor.h"
]
```

- Selected path will be the fbcode one - that's the key to the perf data.
- We want the xplat version to point to the data attached to the fbcode file as well.

Differential Revision: D86834568

fbshipit-source-id: 9afac6b47829120c664a9d9d1dd3241bd45323dc